+++ src/Web/HTML/HTMLVideoElement.js
@@ -63,3 +63,11 @@ exports.setPoster = function (poster) {
     };
   };
 };
+
+exports.srcObject = function (mediaStream) {
+  return function (video) {
+    return function () {
+      video.srcObject = mediaStream;
+    };
+  };
+};