+++ src/Web/HTML/MediaDevices.js
@@ -1,0 +1,9 @@
+"use strict";
+
+exports.getUserMedia_ = function (mediaDevices) {
+  return function (constraints) {
+    return function () {
+      return mediaDevices.getUserMedia(constraints);
+    };
+  };
+};